+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/PolarisDiscoveryProperties.java
@@ -18,6 +18,7 @@
 package com.tencent.cloud.polaris;
 
 import com.tencent.cloud.common.constant.ContextConstant;
+import com.tencent.polaris.api.config.consumer.OutlierDetectionConfig;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -126,6 +127,11 @@ public class PolarisDiscoveryProperties {
 	 * namespace exports if not created.
 	 */
 	private String namespaceExports;
+
+	/**
+	 * Instance detect.
+	 */
+	private Detect detect = new Detect();
 
 	public String getInstanceId() {
 		return instanceId;
@@ -278,6 +284,14 @@ public class PolarisDiscoveryProperties {
 		this.namespaceExports = namespaceExports;
 	}
 
+	public Detect getDetect() {
+		return detect;
+	}
+
+	public void setDetect(Detect detect) {
+		this.detect = detect;
+	}
+
 	@Override
 	public String toString() {
 		return "PolarisDiscoveryProperties{" +
@@ -298,6 +312,29 @@ public class PolarisDiscoveryProperties {
 				", preferIpv6=" + preferIpv6 +
 				", allRecoverEnabled=" + allRecoverEnabled +
 				", namespaceExports='" + namespaceExports + '\'' +
+				", detect=" + detect +
 				'}';
+	}
+
+	public static class Detect {
+		/**
+		 * Enable instance detect or not.
+		 */
+		private OutlierDetectionConfig.When when = OutlierDetectionConfig.When.never;
+
+		public OutlierDetectionConfig.When getWhen() {
+			return when;
+		}
+
+		public void setWhen(OutlierDetectionConfig.When when) {
+			this.when = when;
+		}
+
+		@Override
+		public String toString() {
+			return "Detect{" +
+					"when=" + when +
+					'}';
+		}
 	}
 }