+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - [feat:support config all recover enabled.](https://github.com/Tencent/spring-cloud-tencent/pull/1593)
 - [feat:support stat reporting path aggregation.](https://github.com/Tencent/spring-cloud-tencent/pull/1602)
 - [feat:support namespace exports configuration if not created.](https://github.com/Tencent/spring-cloud-tencent/pull/1611)
+- [feat:support instance detect.](https://github.com/Tencent/spring-cloud-tencent/pull/1616)